LANTERN-BACKEND/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
keras==3.5.0
Pillow==9.0.1
Pillow==10.4.0
torch==2.4.0
